Actualizacion de IdCategorie

+++ app/src/main/java/com/easypick/easypick/viewModels/LocalViewModel.kt
@@ -6,7 +6,7 @@ import com.easypick.easypick.model.Category
 import com.easypick.easypick.model.ItemOrder
 
 class LocalViewModel: ViewModel() {
-    var catSelect : Int = 0
+    var catSelect : Long = 0
     var nameCatSelect: String = ""
     var idStore: Long = 0
     var productosSeleccionados = ArrayList<ItemOrder>()
